refactor(features): split _features.py into one file per extractor
Replace the monolithic _features.py with focused modules.
Public API through pyvisim/features/__init__.py is unchanged.
