+++ pyvisim/features/__init__.py
@@ -1,10 +1,8 @@
-from ._features import (
-    SIFT,
-    DeepConvFeature,
-    Lambda,
-    RootSIFT,
-    feature_extractor_from_dict,
-)
+from ._deep_conv_feature import DeepConvFeature
+from ._lambda import Lambda
+from ._registry import feature_extractor_from_dict
+from ._root_sift import RootSIFT
+from ._sift import SIFT
 
 __all__ = [
     "SIFT",
